+++ tools/modules/update.sh
@@ -78,7 +78,10 @@ if [[ -z "$1" || "$1" == "--build_only" ]]; then
     mkdir -p "$vbuild_dir"
     trap "rm -rf '$LLVM_REPO/$vbuild_dir'" ERR INT
     cd "$vbuild_dir"
-    CXX=$(echo "${BAZEL_CC}" | sed -E 's/(cc)?(-.*)?$/++\2/')
+    CXX=$(basename "${BAZEL_CC}" | sed -E 's/(cc)?(-.*)?$/++\2/')
+    if [ ! -z $(dirname "${BAZEL_CC}") ]; then
+      CXX="$(dirname "${BAZEL_CC}")/${CXX}"
+    fi
     ../configure CC="${BAZEL_C_COMPILER}" CXX="${CXX}" \
       --prefix="$LLVM_REPO/build-install" \
       CXXFLAGS="-std=c++11" \
@@ -86,6 +89,7 @@ if [[ -z "$1" || "$1" == "--build_only" ]]; then
     make -j8
     cd ..
   fi
+  rm -f build
   if [[ $(uname) == 'Darwin' ]]; then
     ln -sf "$vbuild_dir" build
   else